backend/requirements.txt
asgiref
Django
django-anymail
django-ckeditor
django-ckeditor-5
django-cors-headers
django-dotenv
django-environ
django-import-export
django-jazzmin
django-rest-framework
djangorestframework
djangorestframework-simplejwt
drf-yasg
humanize
requests
shortuuid
sqlparse
tzdata